requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
qrcode
pillow
